Update for last semester at Navarro CIT
I plan on doing weekly or biweekly updates to the website from here on out with a new direction and focus. I will be incorporating more JavaScript because all of the HTML and CSS is finished. All that needs to be done is to make it more interactive via JavaScript.